safari-534.53-branch/LayoutTests/plugins/keyboard-events-expected.txt

CONSOLE MESSAGE: line 0: PLUGIN: getFocusEvent
CONSOLE MESSAGE: line 0: PLUGIN: mouseDown at (12, 12)
CONSOLE MESSAGE: line 0: PLUGIN: mouseUp at (12, 12)
CONSOLE MESSAGE: line 0: PLUGIN: keyDown 'a'
CONSOLE MESSAGE: line 0: PLUGIN: keyUp 'a'
CONSOLE MESSAGE: line 0: PLUGIN: keyDown 'b'
CONSOLE MESSAGE: line 0: PLUGIN: keyUp 'b'
CONSOLE MESSAGE: line 0: PLUGIN: keyDown 'c'
CONSOLE MESSAGE: line 0: PLUGIN: keyUp 'c'

This test checks if a plug-in can receive keyboard events sent from eventSender. This is a test for Bug 34936.
